Refacto: split render2.c into more structured files
